py1/requirements.txt
antlr4-python3-runtime==4.9.2
type-docopt
schema